Rename `pimm.methods` to `pimm.calls`
