+++ pimm/__init__.py
@@ -1,4 +1,4 @@
-from . import methods, shared_memory
+from . import calls, shared_memory
 from .core import (
     Clock,
     Command,
@@ -32,9 +32,9 @@ __all__ = [
     'EmitterDict',
     'FakeEmitter',
     'FakeReceiver',
+    'calls',
     'map',
     'Message',
-    'methods',
     'NoOpEmitter',
     'NoOpReceiver',
     'NoValueException',
